refactor: moving modal hook to context
